refactor: update package structure to enhance project organization
- Update Java package names and structure across various classes to better align with clean architecture principles.
