+++ backend/src/main/java/com/harmonical/backend/HarmonicalBackendApplication.java
@@ -1,4 +1,4 @@
-package com.harmonical;
+package com.harmonical.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
