+++ backend/src/main/java/com/harmonical/backend/infrastructure/config/PersistenceConfig.java
@@ -1,4 +1,4 @@
-package com.harmonical.infrastructure.config;
+package com.harmonical.backend.infrastructure.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
-@EnableJpaRepositories(basePackages = "com.harmonical.infrastructure.persistence.repository")
+@EnableJpaRepositories(basePackages = "com.harmonical.backend.infrastructure.persistence.repository")
 @EnableJpaAuditing
 @EnableTransactionManagement
 public class PersistenceConfig {
