+++ backend/src/main/java/com/harmonical/backend/infrastructure/persistence/repository/JPAEventRepository.java
@@ -1,6 +1,6 @@
-package com.harmonical.infrastructure.persistence.repository;
+package com.harmonical.backend.infrastructure.persistence.repository;
 
-import com.harmonical.infrastructure.persistence.model.EventEntity;
+import com.harmonical.backend.infrastructure.persistence.model.EventEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
